Restrict visibility of libraries that are newly vendor_available: true.
As part of fixing b/68319577, some libraries were made vendor_available: true.
To minimize new dependencies from vendor on these libraries, this change
restricts the visibility of the libraries to only the framework projects that
currently use them.

Additionally, the only user of the libbrotli shared library is brotli itself,
so we eliminate the shared library, preventing the shared library from
entering the vndk. The brotli executable now links to the static libbrotli.

Bug: 141649008
Test: Targets build successfully.
Change-Id: I5b9ebdaa44214598f36540cd0dae023d5bf11e42